Refactor sessin info
