+++ app/javascript/react/components/Modals/SessionDetailsModal/SessionInfo/DesktopHeader.tsx
@@ -1,0 +1,149 @@
+import React from "react";
+import { useTranslation } from "react-i18next";
+import calendar from "../../../../assets/icons/calendar.svg";
+import downloadImage from "../../../../assets/icons/download.svg";
+import shareLink from "../../../../assets/icons/shareLink.svg";
+import { white } from "../../../../assets/styles/colors";
+import { MobileStreamShortInfo as StreamShortInfo } from "../../../../types/mobileStream";
+import { Thresholds } from "../../../../types/thresholds";
+import { copyCurrentURL } from "../../../../utils/copyCurrentUrl";
+import { isNoData } from "../../../../utils/measurementsCalc";
+import { getColorForValue } from "../../../../utils/thresholdColors";
+import { ExportDataModal } from "../../../Modals/ExportDataModal";
+import { CopyLinkModal } from "../../CopyLinkModal";
+import * as S from "../SessionDetailsModal.style";
+
+interface Extremes {
+  minMeasurementValue: number | null;
+  maxMeasurementValue: number | null;
+  averageValue: number | null;
+}
+
+interface DesktopHeaderProps {
+  streamShortInfo: StreamShortInfo;
+  thresholds: Thresholds;
+  extremes: Extremes;
+  formattedTime: (time: string) => string;
+  streamId: number | null;
+  fixedSessionTypeSelected: boolean;
+}
+
+const DesktopHeader: React.FC<DesktopHeaderProps> = ({
+  streamShortInfo,
+  thresholds,
+  extremes,
+  formattedTime,
+  streamId,
+  fixedSessionTypeSelected,
+}) => {
+  const { t } = useTranslation();
+
+  const { minMeasurementValue, maxMeasurementValue, averageValue } = extremes;
+  const noData = isNoData(
+    extremes.minMeasurementValue,
+    extremes.maxMeasurementValue,
+    extremes.averageValue
+  );
+  return (
+    <S.DesktopHeader>
+      <S.Wrapper>
+        <S.SessionName>{streamShortInfo.title}</S.SessionName>
+        <S.ProfileName>{streamShortInfo.profile}</S.ProfileName>
+        <S.SensorName>{streamShortInfo.sensorName}</S.SensorName>
+      </S.Wrapper>
+      <S.Wrapper>
+        {noData ? (
+          <S.NoData>{t("sessionDetailsModal.noData")}</S.NoData>
+        ) : (
+          <>
+            <S.AverageValueContainer>
+              <S.AverageDot
+                $color={getColorForValue(thresholds, averageValue)}
+              />
+              {t("sessionDetailsModal.averageValue")}
+              <S.AverageValue>{averageValue}</S.AverageValue>
+              {streamShortInfo.unitSymbol}
+            </S.AverageValueContainer>
+            <S.MinMaxValueContainer $isMobile={false}>
+              <div>
+                <S.SmallDot
+                  $color={getColorForValue(thresholds, minMeasurementValue)}
+                />
+                {t("sessionDetailsModal.minValue")}
+                <S.Value>{minMeasurementValue}</S.Value>
+              </div>
+              <div>
+                <S.SmallDot
+                  $color={getColorForValue(thresholds, maxMeasurementValue)}
+                />
+                {t("sessionDetailsModal.maxValue")}
+                <S.Value>{maxMeasurementValue}</S.Value>
+              </div>
+            </S.MinMaxValueContainer>
+          </>
+        )}
+        <S.TimeRange>
+          {formattedTime(streamShortInfo.startTime ?? "")} -{" "}
+          {formattedTime(streamShortInfo.endTime ?? "")}
+        </S.TimeRange>
+      </S.Wrapper>
+      <S.ButtonsContainer>
+        {fixedSessionTypeSelected && (
+          <S.BlueButton to={`/fixed_stream?streamId=${streamId}`}>
+            {t("sessionDetailsModal.calendar")}
+            <img src={calendar} alt={t("sessionDetailsModal.calendarIcon")} />
+          </S.BlueButton>
+        )}
+        <S.SmallPopup
+          trigger={
+            <S.Button aria-labelledby={t("calendarHeader.altExportSession")}>
+              <img
+                src={downloadImage}
+                alt={t("calendarHeader.altExportSession")}
+              />
+            </S.Button>
+          }
+          position="top center"
+          nested
+          closeOnDocumentClick
+          offsetX={fixedSessionTypeSelected ? 0 : 40}
+          arrowStyle={
+            fixedSessionTypeSelected
+              ? {}
+              : {
+                  left: "34%",
+                  borderColor: `transparent transparent ${white} transparent`,
+                  borderWidth: "0 10px 10px 10px",
+                  borderStyle: "solid",
+                }
+          }
+        >
+          <ExportDataModal
+            sessionId={streamShortInfo.sessionId}
+            onSubmit={(formData) => {}}
+          />
+        </S.SmallPopup>
+        <S.SmallPopup
+          trigger={
+            <S.Button
+              onClick={copyCurrentURL}
+              aria-label={t("copyLinkModal.altCopyLink")}
+            >
+              <img src={shareLink} alt={t("copyLinkModal.copyLink")} />
+            </S.Button>
+          }
+          position="top center"
+          nested
+          closeOnDocumentClick
+        >
+          <CopyLinkModal
+            sessionId={streamShortInfo.sessionId}
+            onSubmit={(formData) => {}}
+          />
+        </S.SmallPopup>
+      </S.ButtonsContainer>
+    </S.DesktopHeader>
+  );
+};
+
+export default DesktopHeader;
